added password field to user and added comments

+++ controllers/userController.js
@@ -1,11 +1,18 @@
 const User = require('../models').User
-const userExists = require('../helpers').resourceExists
+const helper = require('../helpers')
+
 const NOT_FOUND = { message: 'user not found' }
 const SUCCESS = { message: "success" }
 
+/**
+ * create user
+ * delete the password form the object
+ * if unique email error send 409 status
+ */
 const createUser = async (req, res) => {
     try {
-        const user = await User.create(req.body)
+        const user = await User.create(req.body, { raw: true })
+        delete user.dataValues.password
         res.status(201).send(user)
     } catch (error) {
         let status = error.name === "SequelizeUniqueConstraintError" ? 409 : 500
@@ -15,7 +22,7 @@ const createUser = async (req, res) => {
 
 const getAllUsers = async (req, res) => {
     try {
-        const user = await User.findAll()
+        const user = await User.findAll({ attributes: { exclude: ['password'] } })
         res.status(200).send(user)
     } catch (error) {
         res.status(500).send(error)
@@ -26,7 +33,10 @@ const getUser = async (req, res) => {
     let { id } = req.params
 
     try {
-        const user = await User.findByPk(id)
+        let exists = await helper.resourceExists(User, id)
+        if (!exists) return res.status(404).send(NOT_FOUND)
+
+        const user = await User.findByPk(id, { attributes: { exclude: ['password'] } })
         res.status(200).send(user)
     } catch (error) {
         res.status(500).send(error)
@@ -36,11 +46,8 @@ const getUser = async (req, res) => {
 const updateUser = async (req, res) => {
     let { id } = req.params
     try {
-        let exists = await userExists(User, id)
-
-        if (!exists) {
-           return res.status(404).send(NOT_FOUND)
-        }
+        let exists = await helper.resourceExists(User, id)
+        if (!exists) return res.status(404).send(NOT_FOUND)
 
         await User.update(req.body, { where: { id: id } })
         res.status(200).send(SUCCESS)
@@ -52,6 +59,9 @@ const updateUser = async (req, res) => {
 const deleteUser = async (req, res) => {
     let { id } = req.params
     try {
+        let exists = await helper.resourceExists(User, id)
+        if (!exists) return res.status(404).send(NOT_FOUND)
+        
         await User.destroy({ where: { id: id } })
         res.status(200).send(SUCCESS)
     } catch (error) {
